third_party/blink/web_tests/external/wpt/bluetooth/idl/idlharness.tentative.https.window-expected.txt

This is a testharness.js-based test.
Found 51 FAIL, 0 TIMEOUT, 0 NOTRUN.
[FAIL] Bluetooth interface: attribute onavailabilitychanged
  assert_true: The prototype object must have a property "onavailabilitychanged" expected true got false
[FAIL] Bluetooth interface: attribute referringDevice
  assert_true: The prototype object must have a property "referringDevice" expected true got false
[FAIL] Bluetooth interface: attribute ongattserverdisconnected
  assert_true: The prototype object must have a property "ongattserverdisconnected" expected true got false
[FAIL] Bluetooth interface: attribute oncharacteristicvaluechanged
  assert_true: The prototype object must have a property "oncharacteristicvaluechanged" expected true got false
[FAIL] Bluetooth interface: attribute onserviceadded
  assert_true: The prototype object must have a property "onserviceadded" expected true got false
[FAIL] Bluetooth interface: attribute onservicechanged
  assert_true: The prototype object must have a property "onservicechanged" expected true got false
[FAIL] Bluetooth interface: attribute onserviceremoved
  assert_true: The prototype object must have a property "onserviceremoved" expected true got false
[FAIL] Bluetooth interface: navigator.bluetooth must inherit property "onavailabilitychanged" with the proper type
  assert_inherits: property "onavailabilitychanged" not found in prototype chain
[FAIL] Bluetooth interface: navigator.bluetooth must inherit property "referringDevice" with the proper type
  assert_inherits: property "referringDevice" not found in prototype chain
[FAIL] Bluetooth interface: navigator.bluetooth must inherit property "ongattserverdisconnected" with the proper type
  assert_inherits: property "ongattserverdisconnected" not found in prototype chain
[FAIL] Bluetooth interface: navigator.bluetooth must inherit property "oncharacteristicvaluechanged" with the proper type
  assert_inherits: property "oncharacteristicvaluechanged" not found in prototype chain
[FAIL] Bluetooth interface: navigator.bluetooth must inherit property "onserviceadded" with the proper type
  assert_inherits: property "onserviceadded" not found in prototype chain
[FAIL] Bluetooth interface: navigator.bluetooth must inherit property "onservicechanged" with the proper type
  assert_inherits: property "onservicechanged" not found in prototype chain
[FAIL] Bluetooth interface: navigator.bluetooth must inherit property "onserviceremoved" with the proper type
  assert_inherits: property "onserviceremoved" not found in prototype chain
[FAIL] BluetoothPermissionResult interface: existence and properties of interface object
  assert_own_property: self does not have own property "BluetoothPermissionResult" expected property "BluetoothPermissionResult" missing
[FAIL] BluetoothPermissionResult interface object length
  assert_own_property: self does not have own property "BluetoothPermissionResult" expected property "BluetoothPermissionResult" missing
[FAIL] BluetoothPermissionResult interface object name
  assert_own_property: self does not have own property "BluetoothPermissionResult" expected property "BluetoothPermissionResult" missing
[FAIL] BluetoothPermissionResult interface: existence and properties of interface prototype object
  assert_own_property: self does not have own property "BluetoothPermissionResult" expected property "BluetoothPermissionResult" missing
[FAIL] BluetoothPermissionResult interface: existence and properties of interface prototype object's "constructor" property
  assert_own_property: self does not have own property "BluetoothPermissionResult" expected property "BluetoothPermissionResult" missing
[FAIL] BluetoothPermissionResult interface: existence and properties of interface prototype object's @@unscopables property
  assert_own_property: self does not have own property "BluetoothPermissionResult" expected property "BluetoothPermissionResult" missing
[FAIL] BluetoothPermissionResult interface: attribute devices
  assert_own_property: self does not have own property "BluetoothPermissionResult" expected property "BluetoothPermissionResult" missing
[FAIL] ValueEvent interface: existence and properties of interface object
  assert_own_property: self does not have own property "ValueEvent" expected property "ValueEvent" missing
[FAIL] ValueEvent interface object length
  assert_own_property: self does not have own property "ValueEvent" expected property "ValueEvent" missing
[FAIL] ValueEvent interface object name
  assert_own_property: self does not have own property "ValueEvent" expected property "ValueEvent" missing
[FAIL] ValueEvent interface: existence and properties of interface prototype object
  assert_own_property: self does not have own property "ValueEvent" expected property "ValueEvent" missing
[FAIL] ValueEvent interface: existence and properties of interface prototype object's "constructor" property
  assert_own_property: self does not have own property "ValueEvent" expected property "ValueEvent" missing
[FAIL] ValueEvent interface: existence and properties of interface prototype object's @@unscopables property
  assert_own_property: self does not have own property "ValueEvent" expected property "ValueEvent" missing
[FAIL] ValueEvent interface: attribute value
  assert_own_property: self does not have own property "ValueEvent" expected property "ValueEvent" missing
[FAIL] BluetoothDevice interface: attribute oncharacteristicvaluechanged
  assert_true: The prototype object must have a property "oncharacteristicvaluechanged" expected true got false
[FAIL] BluetoothDevice interface: attribute onserviceadded
  assert_true: The prototype object must have a property "onserviceadded" expected true got false
[FAIL] BluetoothDevice interface: attribute onservicechanged
  assert_true: The prototype object must have a property "onservicechanged" expected true got false
[FAIL] BluetoothDevice interface: attribute onserviceremoved
  assert_true: The prototype object must have a property "onserviceremoved" expected true got false
[FAIL] BluetoothAdvertisingEvent interface object length
  assert_equals: wrong value for BluetoothAdvertisingEvent.length expected 2 but got 0
[FAIL] BluetoothAdvertisingEvent must be primary interface of event
  assert_equals: wrong typeof object expected "object" but got "undefined"
[FAIL] Stringification of event
  assert_equals: wrong typeof object expected "object" but got "undefined"
[FAIL] BluetoothAdvertisingEvent interface: event must inherit property "device" with the proper type
  assert_equals: wrong typeof object expected "object" but got "undefined"
[FAIL] BluetoothAdvertisingEvent interface: event must inherit property "uuids" with the proper type
  assert_equals: wrong typeof object expected "object" but got "undefined"
[FAIL] BluetoothAdvertisingEvent interface: event must inherit property "name" with the proper type
  assert_equals: wrong typeof object expected "object" but got "undefined"
[FAIL] BluetoothAdvertisingEvent interface: event must inherit property "appearance" with the proper type
  assert_equals: wrong typeof object expected "object" but got "undefined"
[FAIL] BluetoothAdvertisingEvent interface: event must inherit property "txPower" with the proper type
  assert_equals: wrong typeof object expected "object" but got "undefined"
[FAIL] BluetoothAdvertisingEvent interface: event must inherit property "rssi" with the proper type
  assert_equals: wrong typeof object expected "object" but got "undefined"
[FAIL] BluetoothAdvertisingEvent interface: event must inherit property "manufacturerData" with the proper type
  assert_equals: wrong typeof object expected "object" but got "undefined"
[FAIL] BluetoothAdvertisingEvent interface: event must inherit property "serviceData" with the proper type
  assert_equals: wrong typeof object expected "object" but got "undefined"
[FAIL] BluetoothRemoteGATTService interface: existence and properties of interface object
  assert_equals: prototype of BluetoothRemoteGATTService is not EventTarget expected function "function EventTarget() { [native code] }" but got function "function () { [native code] }"
[FAIL] BluetoothRemoteGATTService interface: existence and properties of interface prototype object
  assert_equals: prototype of BluetoothRemoteGATTService.prototype is not EventTarget.prototype expected object "[object EventTarget]" but got object "[object Object]"
[FAIL] BluetoothRemoteGATTService interface: operation getIncludedService(BluetoothServiceUUID)
  assert_own_property: interface prototype object missing non-static operation expected property "getIncludedService" missing
[FAIL] BluetoothRemoteGATTService interface: operation getIncludedServices(optional BluetoothServiceUUID)
  assert_own_property: interface prototype object missing non-static operation expected property "getIncludedServices" missing
[FAIL] BluetoothRemoteGATTService interface: attribute oncharacteristicvaluechanged
  assert_true: The prototype object must have a property "oncharacteristicvaluechanged" expected true got false
[FAIL] BluetoothRemoteGATTService interface: attribute onserviceadded
  assert_true: The prototype object must have a property "onserviceadded" expected true got false
[FAIL] BluetoothRemoteGATTService interface: attribute onservicechanged
  assert_true: The prototype object must have a property "onservicechanged" expected true got false
[FAIL] BluetoothRemoteGATTService interface: attribute onserviceremoved
  assert_true: The prototype object must have a property "onserviceremoved" expected true got false
Harness: the test ran to completion.
